chore: restructure project for v1.0 with web and CLI versions
      - Move CLI v0 files to cli/ directory for isolation
      - Create root go.mod for server/pkg modules
      - Update README.md to document both web (v1.0) and CLI (v0) versions
      - Create cli/README.md with complete CLI documentation
      - Keep shared resources (locations/, weather_codes/) in root
      - Maintain backward compatibility via cli-v0 branch

      Breaking changes:
      - CLI users must now run from cli/ directory: `cd cli && go run .`
      - Root go.mod module changed to support web version structure

      Project now supports:
      - Web version (default): Go server + Next.js frontend
      - CLI version: Simple command-line tool in cli/ directory
      chore: restructure project for v1.0 with web and CLI versions

     - Move CLI v0 files to cli/ directory for isolation
     - Create root go.mod for server/pkg modules
     - Update README.md to document both web (v1.0) and CLI (v0) versions
     - Create cli/README.md with complete CLI documentation
     - Keep shared resources (locations/, weather_codes/) in root
     - Maintain backward compatibility via cli-v0 branch

     Breaking changes:
     - CLI users must now run from cli/ directory: `cd cli && go run .`
     - Root go.mod module changed to support web version structure

     Project now supports:
     - Web version (default): Go server + Next.js frontend
     - CLI version: Simple command-line tool in cli/ directory
