+++ cli/returnFormat.json
@@ -1,0 +1,25 @@
+{
+    "latitude": 46.94,
+    "longitude": 7.44,
+    "generationtime_ms": 0.06639957427978516,
+    "utc_offset_seconds": 0,
+    "timezone": "GMT",
+    "timezone_abbreviation": "GMT",
+    "elevation": 554.0,
+    "current_units": {
+        "time": "iso8601",
+        "interval": "seconds",
+        "temperature_2m": "°C",
+        "relative_humidity_2m": "%",
+        "rain": "mm",
+        "weather_code": "wmo code"
+    },
+    "current": {
+        "time": "2025-08-08T07:15",
+        "interval": 900,
+        "temperature_2m": 18.3,
+        "relative_humidity_2m": 80,
+        "rain": 0.00,
+        "weather_code": 0
+    }
+}
